Revert "feat(app-control): switch observe to a continuous SCStream per window (#29445)"
This reverts commit 09c9bb1481b6b5e3398d620db8f247c90c190a05.
